Refactor triage analyzers (#1005)
* triage analyzer refactor to remove code duplication

* triage analyzer refactor to remove code duplication

* triage analyzer refactor to remove code duplication

* triage analyzer refactor to remove code duplication

* triage analyzer refactor to remove code duplication

* adjust

* fixes
